amamessage_deploy_production_20250726_232622/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.12.1
pydantic==2.5.0
pydantic-settings==2.1.0
python-multipart==0.0.6
jinja2==3.1.2
aiofiles==23.2.1
httpx==0.25.2
python-dotenv==1.0.0
bcrypt==4.1.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
schedule==1.2.0
asyncpg==0.29.0
celery==5.3.4
redis==5.0.1
pyserial==3.5
requests==2.31.0
